refactor(backend): collapse two crates into one
