+++ Cargo.toml
@@ -43,7 +43,6 @@ members = [
     "crates/utils/dependent",
     "crates/utils/external",
     "crates/utils/env",
-    "crates/utils/specific",
 ]
 resolver = "2"
