+++ crates/services/importer/src/audiobookshelf.rs
@@ -9,6 +9,7 @@ use data_encoding::BASE64;
 use dependent_models::{ImportCompletedItem, ImportResult};
 use dependent_utils::get_identifier_from_book_isbn;
 use enum_models::{ImportSource, MediaLot, MediaSource};
+use external_utils::audiobookshelf::get_updated_metadata;
 use media_models::{
     DeployUrlAndKeyImportInput, ImportOrExportMetadataItem, ImportOrExportMetadataItemSeen,
     UniqueMediaIdentifier,
@@ -20,7 +21,6 @@ use reqwest::{
 };
 use serde_json::json;
 use specific_models::audiobookshelf as audiobookshelf_models;
-use specific_utils::audiobookshelf::get_updated_metadata;
 use supporting_service::SupportingService;
 
 use super::{ImportFailStep, ImportFailedItem};
