+++ crates/services/integration/Cargo.toml
@@ -32,7 +32,6 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 sonarr-api-rs = { workspace = true }
 specific-models = { path = "../../models/specific" }
-specific-utils = { path = "../../utils/specific" }
 supporting-service = { path = "../supporting" }
 tokio = { workspace = true }
 tracing = { workspace = true }
