+++ crates/utils/external/src/lib.rs
@@ -1,11 +1,20 @@
+use std::sync::Arc;
+
 use anyhow::Result;
-use common_utils::{ryot_log, APPLICATION_JSON_HEADER, USER_AGENT_STR};
+use common_utils::{ryot_log, sleep_for_n_seconds, APPLICATION_JSON_HEADER, USER_AGENT_STR};
+use database_models::{metadata, prelude::Metadata};
+use dependent_utils::deploy_update_metadata_job;
+use enum_models::{MediaLot, MediaSource};
 use reqwest::{
     header::{HeaderMap, HeaderValue, ACCEPT, AUTHORIZATION, USER_AGENT},
     Client, ClientBuilder,
 };
-use sea_orm::prelude::DateTimeUtc;
+use sea_orm::{
+    prelude::DateTimeUtc,
+    {ColumnTrait, EntityTrait, QueryFilter},
+};
 use serde::{Deserialize, Serialize};
+use supporting_service::SupportingService;
 
 pub mod jellyfin {
     use super::*;
@@ -105,5 +114,35 @@ pub mod jellyfin {
         ryot_log!(debug, "Authenticated as user id: {}", user_id);
 
         Ok((client, user_id))
+    }
+}
+
+pub mod audiobookshelf {
+    use super::*;
+
+    pub async fn get_updated_metadata(
+        identifier: &String,
+        ss: &Arc<SupportingService>,
+    ) -> Result<metadata::Model> {
+        async fn get_metadata(
+            identifier: &String,
+            ss: &Arc<SupportingService>,
+        ) -> Result<metadata::Model> {
+            let m = Metadata::find()
+                .filter(metadata::Column::Identifier.eq(identifier))
+                .filter(metadata::Column::Lot.eq(MediaLot::Podcast))
+                .filter(metadata::Column::Source.eq(MediaSource::Itunes))
+                .one(&ss.db)
+                .await?
+                .unwrap();
+            Ok(m)
+        }
+
+        let already = get_metadata(identifier, ss).await?;
+        if already.podcast_specifics.is_none() {
+            deploy_update_metadata_job(&already.id, ss).await.unwrap();
+            sleep_for_n_seconds(3).await;
+        }
+        get_metadata(identifier, ss).await
     }
 }
